backend/requirements.txt
asgiref==3.11.1
Django==6.0.2

django-filter==25.1
djangorestframework==3.16.1
psycopg2==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
